Really fix the build and make it work. Skip libjuce_jni.so. Just load app .so.
It is maybe due to the way how build targets could be parallelize builds.
the primary target does not really reference libEq_Standalone.so, and targets
in Eq probably do not really contain `Eq_Standalone` as a library (I believed
it is), so `juce_jni` would have built without Eq_Standalone. But since
it is explicitly referenced by `juce_jni`, it caused build failure because
it does not exist at that time yet. It is somehow built in order on my
machines (for unknown reason) and did not cause the build failures.

The fix here is to entirely skip libjuce_jni.so anymore, so that loading
only app shared library just works.

To make it happen, we had to make changes to app startup Java.kt to load
app-specific shared library instead. This means, we have to rewrite this
class for each application, but that should be acceptable.

In case that interferes with JUCE core code in the future, we can introduce
another startup provider that is dedicated to the app so loading.